# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type AccountType = (typeof AccountType)[keyof typeof AccountType];

export const AccountType = {
  checking: "checking",
  savings: "savings",
  credit: "credit",
  investment: "investment",
  loan: "loan",
  cash: "cash",
  other: "other",
} as const;

export interface Account {
  id: number;
  name: string;
  type: AccountType;
  balance: number;
  currency: string;
  color?: string;
  icon?: string;
  institution?: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateAccountRequestType =
  (typeof CreateAccountRequestType)[keyof typeof CreateAccountRequestType];

export const CreateAccountRequestType = {
  checking: "checking",
  savings: "savings",
  credit: "credit",
  investment: "investment",
  loan: "loan",
  cash: "cash",
  other: "other",
} as const;

export interface CreateAccountRequest {
  name: string;
  type: CreateAccountRequestType;
  balance: number;
  currency?: string;
  color?: string;
  icon?: string;
  institution?: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType];

export const CategoryType = {
  income: "income",
  expense: "expense",
  both: "both",
} as const;

export interface Category {
  id: number;
  name: string;
  type: CategoryType;
  color: string;
  icon: string;
  parentId?: number;
  isDefault: boolean;
  createdAt: string;
}

export interface Transaction {
  id: number;
  accountId: number;
  categoryId?: number;
  type: TransactionType;
  amount: number;
  description: string;
  notes?: string;
  date: string;
  isRecurring: boolean;
  recurringInterval?: string;
  tags?: string;
  createdAt: string;
  updatedAt: string;
  account?: Account;
  category?: Category;
}

export type CreateTransactionRequestType =
  (typeof CreateTransactionRequestType)[keyof typeof CreateTransactionRequestType];

export const CreateTransactionRequestType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export interface CreateTransactionRequest {
  accountId: number;
  categoryId?: number;
  type: CreateTransactionRequestType;
  amount: number;
  description: string;
  notes?: string;
  date: string;
  isRecurring?: boolean;
  recurringInterval?: string;
  tags?: string;
}

export interface TransactionListResponse {
  data: Transaction[];
  total: number;
  limit: number;
  offset: number;
}

export type CreateCategoryRequestType =
  (typeof CreateCategoryRequestType)[keyof typeof CreateCategoryRequestType];

export const CreateCategoryRequestType = {
  income: "income",
  expense: "expense",
  both: "both",
} as const;

export interface CreateCategoryRequest {
  name: string;
  type: CreateCategoryRequestType;
  color: string;
  icon: string;
  parentId?: number;
}

export interface Budget {
  id: number;
  categoryId: number;
  amount: number;
  spent: number;
  month: number;
  year: number;
  alertThreshold?: number;
  createdAt: string;
  category?: Category;
}

export interface CreateBudgetRequest {
  categoryId: number;
  amount: number;
  month: number;
  year: number;
  alertThreshold?: number;
}

export type GoalStatus = (typeof GoalStatus)[keyof typeof GoalStatus];

export const GoalStatus = {
  active: "active",
  completed: "completed",
  paused: "paused",
} as const;

export interface Goal {
  id: number;
  name: string;
  targetAmount: number;
  currentAmount: number;
  targetDate?: string;
  color?: string;
  icon?: string;
  description?: string;
  status: GoalStatus;
  createdAt: string;
  updatedAt: string;
}

export type CreateGoalRequestStatus =
  (typeof CreateGoalRequestStatus)[keyof typeof CreateGoalRequestStatus];

export const CreateGoalRequestStatus = {
  active: "active",
  completed: "completed",
  paused: "paused",
} as const;

export interface CreateGoalRequest {
  name: string;
  targetAmount: number;
  currentAmount?: number;
  targetDate?: string;
  color?: string;
  icon?: string;
  description?: string;
  status?: CreateGoalRequestStatus;
}

export interface AnalyticsSummary {
  totalIncome: number;
  totalExpenses: number;
  netSavings: number;
  savingsRate: number;
  transactionCount: number;
  topCategory?: string;
  periodLabel: string;
}

export interface CategorySpending {
  categoryId: number;
  categoryName: string;
  color: string;
  icon: string;
  amount: number;
  percentage: number;
  transactionCount: number;
}

export interface MonthlyTrend {
  month: string;
  income: number;
  expenses: number;
  savings: number;
}

export interface NetWorth {
  totalAssets: number;
  totalLiabilities: number;
  netWorth: number;
  accounts: Account[];
}

export type ListTransactionsParams = {
  accountId?: number;
  categoryId?: number;
  type?: ListTransactionsType;
  startDate?: string;
  endDate?: string;
  search?: string;
  limit?: number;
  offset?: number;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export type ListBudgetsParams = {
  month?: number;
  year?: number;
};

export type GetAnalyticsSummaryParams = {
  month?: number;
  year?: number;
};

export type GetSpendingByCategoryParams = {
  month?: number;
  year?: number;
};

export type GetMonthlyTrendsParams = {
  months?: number;
};