# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Personal Finance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get an account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get financial summary
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
  /analytics/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [analytics]
      summary: Get spending grouped by category
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpending"
  /analytics/monthly-trends:
    get:
      operationId: getMonthlyTrends
      tags: [analytics]
      summary: Get monthly income vs expense trends
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Monthly trends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrend"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, credit, investment, loan, cash, other]
        balance:
          type: number
        currency:
          type: string
        color:
          type: string
        icon:
          type: string
        institution:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - type
        - balance
        - currency
        - isActive
        - createdAt
        - updatedAt
    CreateAccountRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, credit, investment, loan, cash, other]
        balance:
          type: number
        currency:
          type: string
        color:
          type: string
        icon:
          type: string
        institution:
          type: string
      required:
        - name
        - type
        - balance
...
    AnalyticsSummary:
      type: object
      properties:
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netSavings:
          type: number
        savingsRate:
          type: number
        transactionCount:
          type: integer
        topCategory:
          type: string
        periodLabel:
          type: string
      required:
        - totalIncome
        - totalExpenses
        - netSavings
        - savingsRate
        - transactionCount
        - periodLabel
    CategorySpending:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        color:
          type: string
        icon:
          type: string
        amount:
          type: number
        percentage:
          type: number
        transactionCount:
          type: integer
      required:
        - categoryId
        - categoryName
        - color
        - icon
        - amount
        - percentage
        - transactionCount
    MonthlyTrend:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        expenses:
          type: number
        savings:
          type: number
      required:
        - month
        - income
        - expenses
        - savings
...